requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pytest
pytest-mock
httpx
cryptography
pytest 
pytest-asyncio 
httpx